BlockEventListener: implement handler for `BlockExplodeEvent`
