+++ src/main/java/de/lmichaelis/aurora/listener/EntityEventListener.java
@@ -3,6 +3,7 @@
 package de.lmichaelis.aurora.listener;
 
 import de.lmichaelis.aurora.Aurora;
+import de.lmichaelis.aurora.AuroraUtil;
 import de.lmichaelis.aurora.Predicates;
 import de.lmichaelis.aurora.model.Claim;
 import de.lmichaelis.aurora.model.Group;
@@ -140,25 +141,7 @@ public final class EntityEventListener extends BaseListener {
 
 	@EventHandler(ignoreCancelled = true)
 	public void onEntityExplode(final @NotNull EntityExplodeEvent event) {
-		final var iter = event.blockList().iterator();
-		while (iter.hasNext()) {
-			final var block = iter.next();
-
-			// Ignore air blocks
-			if (block.getType().isAir()) continue;
-
-			// TODO: Cache the claim for better efficiency
-			final var claim = Claim.getClaim(block.getLocation());
-
-			// Rule: Explosions can affect all blocks outside of claims
-			if (claim == null) continue;
-
-			// Rule: If explosions are turned on in the claim, all blocks can be destroyed
-			if (claim.allowsExplosions) continue;
-
-			// Otherwise, prevent the block from breaking
-			iter.remove();
-		}
+		AuroraUtil.neutralizeExplosion(event.blockList());
 	}
 
 	@EventHandler(ignoreCancelled = true)
